Add workflow to run example empress_cli commands (#69)
Add github workflow to run empress on headless mode. The workflow runs the four example empress commands on readme and check whether the commands are successful.

I also plan to have the work flow run lint to check for syntax errors. But when I do that, I see that there are still many syntax errors in the code, so I have not enabled linting in this pull request.