New reconciliation vis package
This package is intended to replace vis_trans and topo_sort.  I recommend removing those branches and merging this one into the master.  This contains the topo_sort functionality in util.py and replaces vis_trans with a new version based on matplotlib rather than tkinter.  render_test1.py and render_test2.py demonstrate the functionality for the trees and reconcililations in sample_data.py.