Add visualization tests and upload output as build artifacts (#91)
Previously this week, Prof Ran told me that the tests I wrote for visualization code might be too hard for people to understand and navigate. After thinking about it, I agree that it is quite odd that a person has to pause in the debugger to see the output of a test. Still, I want the test to be automated. I was thinking about a solution that would both be easy to understand and is also automated.

I think a solution to this is, instead of using unittests, we can write simple scripts for testing visualization code that will write an output to a file. The scripts will be similar to the scripts Prof Ran originally wrote for visualization testing. These scripts will all be in a folder named draw at top-level. The ci will call each script in this folder and will upload the images drawn by the script to what's called the "artifact" of the build.