Remove the display optional argument from CLI cost-regions (#176)
We remove --display from the optional arguments of the cost-regions command of the command-line interface. The command-line interface should not display a pop-up menu. The user can use the GUI if he/she wants a pop-up.

The reason we set the matplotlib backend was that we want the pop-up to work correctly on the command-line interface. Now that we don't display a pop-up anymore, we can remove that line.