Fix tanglegram (#191)
* Changed tanglegram viewer to place tip labels along the branches that lead to the tips
to make the images more readable.  Also, changed recon_viewer.py to offset the labels of the
parasite tips so that they aren't superimposed on the host tips.  This solution is a bit hacky,
but it seems to work reasonably well.

* Updated plot_tools and recon_viewer to draw reconciliations and tanglegrams
more cleanly.

* Fixed issue with where monospace font family was set.  Previously, it was changing rcParam which was a global change to the
font family which is not good!  This was fixed so that only the text_v2 function in plot_tools uses monospace.  Also, got rid
of the title "Reconciliation" in the reconciliation viewer since this is not needed information.

* Changed width of lines in the reconciliation viewer to make the drawings easier to see.  Remove dots at tips also to
make rendering of reconcilation easier to see.

* Updated reconciliation rendering to make it a bit easier to read.

* Addressed Ross's comments in review.

* Removed unused line of code.