Fix reconciliation button graying out bug
As documented in #180, sometimes `View reconciliations` and `View p-value histogram` buttons become unexpectedly grayed out after setting the transfer cost to 1.5,  clicking compute reconciliations, and then set the number of cluster.

This happens because our code says we will gray out the compute reconciliation button and other buttons when an action happens to the three cost boxes (duplication, transfer, loss). The problem with this is that when we click on `View solution space` -> `Clusters`, the event `focusout` and `focusin` are triggered by the input boxes, graying out the button. The fix is to have the gray out command run only on `key` event, which is triggered on text change.

Resolve #180