Allow host tree and parasite tree to have the same node name (#233)
This change allows host tree and parasite tree to have the same node name.

The topological order algorithm has a line that sort the set of vertices of both the host nodes and parasite nodes. This create an error if a host node and a parasite node has the same name. Host node and parasite node is a two-tuple (name: str, nodetype: TreeType) where nodetype is not comparable. When they have the same name, python tries to compare nodetype while sorting, leading to an error. Making nodetype comparable allows host tree and parasite tree to have the same node name.

Resolves #232